refactor: bottomButton icon 컴포넌트 및 타입 분리
