+++ src/screens/screening/detail/components/bottomDetailButton/BottomDetailButton.style.ts
@@ -1,5 +1,37 @@
-import palette from '@/styles/theme/color';
 import {StyleSheet} from 'react-native';
+
+import palette from '@/styles/theme/color';
+
+interface IBottomDetail {
+  text: string;
+  color: string;
+  disabled: boolean;
+  heartDisabled: boolean;
+}
+
+export const bottomDetailTypesStyles: Record<
+  'complete' | 'reviewStart' | 'reviewComplete',
+  IBottomDetail
+> = {
+  complete: {
+    text: '신청완료',
+    color: palette.Another.White,
+    disabled: true,
+    heartDisabled: false,
+  },
+  reviewStart: {
+    text: '리뷰하기',
+    color: palette.Text.Normal,
+    disabled: false,
+    heartDisabled: true,
+  },
+  reviewComplete: {
+    text: '리뷰완료',
+    color: palette.Another.White,
+    disabled: true,
+    heartDisabled: true,
+  },
+};
 
 export const bottomButtonStyles = StyleSheet.create({
   container: {
