+++ src/types/ui.ts
@@ -39,6 +39,14 @@ export type TextInputStyle = {
 // optionButton 타입 지정
 export type OptionButtonType = 'heart' | 'alarm' | 'write';
 
+// 디테일 페이지 BottomDetailButton 타입 지정
+export type DetailBottomButtonType =
+  | 'default'
+  | 'complete'
+  | 'finish'
+  | 'reviewStart'
+  | 'reviewComplete';
+
 // 공통 컴포넌트 중복 prop 타입 지정
 export interface CommonTextProps {
   text: string;
